Prepare for moving image into a new div, topbar
Currently, the protester’s image a 100 x 100 image as CSS. However, it
seems that if those dimensions are adjusted, it leads to text
overlapping. My making a topbar header, we are able to constrain the
image within the title, therefore making the text below it not overlap
with the image.